refactor(encoding/binary,glog/logger): reorganize

+++ pkg/glog/logger/logger.go
@@ -69,7 +69,7 @@ func (this *Logger) genLog(level iface.Level) *Log {
 
 	info := &preInfo{
 		Pkg:   this.pkg,
-		Level: uint8(level),
+		Level: level,
 	}
 	return genLog(this.logger, info)
 }
